requirements.txt
astroid
autopep8
certifi
chardet
click
Flask
gunicorn
idna
isort
itsdangerous
Jinja2
lazy-object-proxy
MarkupSafe
mccabe
numpy
pycodestyle
PyJWT
pylint
python-dotenv
pytz
requests
six
toml
twilio
urllib3
Werkzeug
wrapt
